requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0
pandas==2.2.0
python-dateutil==2.8.2
lxml==4.9.3
playwright==1.57.0
Flask==2.3.4
